+++ src/main/java/de/markusbordihn/playercompanions/entity/type/healer/HealerEntityFlyingAround.java
@@ -19,18 +19,20 @@
 
 package de.markusbordihn.playercompanions.entity.type.healer;
 
+import java.util.Map;
+
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
 import net.minecraft.world.level.Level;
-import java.util.Map;
+
 import de.markusbordihn.playercompanions.entity.AggressionLevel;
 import de.markusbordihn.playercompanions.entity.PlayerCompanionEntity;
 import de.markusbordihn.playercompanions.entity.PlayerCompanionEntityFlying;
 import de.markusbordihn.playercompanions.entity.PlayerCompanionVariant;
 import de.markusbordihn.playercompanions.entity.type.PlayerCompanionType;
+import de.markusbordihn.playercompanions.entity.type.PlayerCompanionTypeIcon;
 
 public class HealerEntityFlyingAround extends PlayerCompanionEntityFlying {
 
@@ -40,12 +42,20 @@ public class HealerEntityFlyingAround extends PlayerCompanionEntityFlying {
       Level level, Map<PlayerCompanionVariant, ResourceLocation> textureByVariant,
       Map<PlayerCompanionVariant, Item> companionItemByVariant) {
     super(entityType, level, textureByVariant, companionItemByVariant);
-    this.setCompanionType(PlayerCompanionType.HEALER);
-    this.setCompanionTypeIcon(new ItemStack(Items.POTION));
     this.setAggressionLevel(AggressionLevel.PASSIVE);
 
     // Shared Healer Features
     this.healerFeatures = new HealerFeatures(this, level);
+  }
+
+  @Override
+  public PlayerCompanionType getCompanionType() {
+    return PlayerCompanionType.HEALER;
+  }
+
+  @Override
+  public ItemStack getCompanionTypeIcon() {
+    return PlayerCompanionTypeIcon.HEALER;
   }
 
   @Override